refactor: usenonfungibleassets
