+++ packages/dashboard/src/pages/Wallets/components/HistoryTable/index.tsx
@@ -2,11 +2,11 @@ import { type Dispatch, memo, type SetStateAction, useMemo, useState } from 'rea
 import { Box, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Typography } from '@mui/material'
 import { LoadingBase, makeStyles, MaskColorVar } from '@masknet/theme'
 import type { Web3Helper } from '@masknet/web3-helpers'
-import { useTransactions, useNetworkContext } from '@masknet/web3-hooks-base'
+import { useTransactions, useNetworkContext, useIterator } from '@masknet/web3-hooks-base'
 import type { Transaction } from '@masknet/web3-shared-base'
 import { HistoryTableRow } from '../HistoryTableRow/index.js'
 import { useDashboardI18N } from '../../../../locales/index.js'
-import { ElementAnchor, useIterator } from '@masknet/shared'
+import { ElementAnchor } from '@masknet/shared'
 import { EMPTY_LIST } from '@masknet/shared-base'
 import { Icons } from '@masknet/icons'
