+++ packages/shared/src/hooks/index.ts
@@ -2,7 +2,6 @@ export * from './useMenu.js'
 export * from './useRemarkable.js'
 export * from './useOpenShareTxDialog.js'
 export * from './useSnackbarCallback.js'
-export * from './useIterator.js'
 export * from './useImageURL.js'
 export * from './useIsImageURL.js'
 export * from './useTokenSecurity.js'
