+++ packages/web3-hooks/base/src/index.ts
@@ -101,3 +101,4 @@ export * from './useWeb3UI.js'
 export * from './useZeroAddress.js'
 export * from './usePageableAsync.js'
 export * from './ScopedDomains/index.js'
+export * from './useIterator.js'
