+++ packages/web3-hooks/base/src/useNonFungibleAssets.ts
@@ -1,20 +1,16 @@
-import { useCallback, useEffect, useMemo, useState } from 'react'
-import { useUpdateEffect } from 'react-use'
+import { useMemo } from 'react'
 import type { Web3Helper } from '@masknet/web3-helpers'
-import { pageableToIterator, flattenAsyncIterator, EMPTY_LIST, type NetworkPluginID } from '@masknet/shared-base'
+import { pageableToIterator, flattenAsyncIterator, type NetworkPluginID, EMPTY_LIST } from '@masknet/shared-base'
 import { useChainContext } from './useContext.js'
 import { useNetworkDescriptors } from './useNetworkDescriptors.js'
 import { useWeb3State } from './useWeb3State.js'
+import { useIterator } from './useIterator.js'
 
 export function useNonFungibleAssets<S extends 'all' | void = void, T extends NetworkPluginID = NetworkPluginID>(
     pluginID?: T,
     schemaType?: Web3Helper.SchemaTypeScope<S, T>,
     options?: Web3Helper.Web3HubOptionsScope<S, T>,
 ) {
-    const [assets, setAssets] = useState<Array<Web3Helper.NonFungibleAssetScope<S, T>>>(EMPTY_LIST)
-    const [done, setDone] = useState(false)
-    const [loading, toggleLoading] = useState(false)
-    const [error, setError] = useState<string>()
     const { Hub } = useWeb3State(pluginID)
     const { account, chainId } = useChainContext({ account: options?.account })
     const networks = useNetworkDescriptors(pluginID)
@@ -23,8 +19,6 @@ export function useNonFungibleAssets<S extends 'all' | void = void, T extends Ne
     const iterator = useMemo(() => {
         const hub = Hub?.getHub?.({ account, chainId, ...options })
         if (!account || !hub?.getNonFungibleAssets) return
-        setAssets(EMPTY_LIST)
-        setDone(false)
 
         return flattenAsyncIterator(
             networks
@@ -41,58 +35,10 @@ export function useNonFungibleAssets<S extends 'all' | void = void, T extends Ne
                 }),
         )
     }, [Hub, account, JSON.stringify(options), networks, chainId])
-
-    const next = useCallback(async () => {
-        if (!iterator || done) return
-        setError(undefined)
-        toggleLoading(true)
-        const batchResult: Array<Web3Helper.NonFungibleAssetScope<S, T>> = []
-        try {
-            for (const _ of Array.from({ length: options?.size ?? 36 })) {
-                const { value, done: iteratorDone } = await iterator.next()
-                if (value instanceof Error) {
-                    // Controlled error
-                    setError(value.message)
-                    break
-                } else {
-                    if (iteratorDone) {
-                        setDone(true)
-                        break
-                    }
-                    if (!iteratorDone && value) {
-                        batchResult.push(value)
-                    }
-                }
-            }
-        } catch (error_) {
-            // Uncontrolled error
-            setError(error_ as string)
-            setDone(true)
-        }
-        setAssets((pred) => {
-            return [...pred, ...batchResult]
-        })
-        toggleLoading(false)
-    }, [iterator, done])
-
-    // Execute once after next update
-    useEffect(() => {
-        if (next) next()
-    }, [next])
-
-    // clear assets after account updated
-    useUpdateEffect(() => {
-        setAssets(EMPTY_LIST)
-    }, [account])
-
-    const retry = useCallback(() => {
-        setError(undefined)
-        setAssets(EMPTY_LIST)
-        setDone(false)
-    }, [])
+    const { value: assets, next, retry, loading, done, error } = useIterator(iterator)
 
     const value = useMemo(() => {
-        return assets.filter((x) => (options?.chainId ? x.chainId === options.chainId : true))
+        return assets?.filter((x) => (options?.chainId ? x.chainId === options.chainId : true)) ?? EMPTY_LIST
     }, [assets, options?.chainId])
 
     return {
